+++ prowler/providers/m365/services/defenderxdr/defenderxdr_endpoint_privileged_user_exposed_credentials/defenderxdr_endpoint_privileged_user_exposed_credentials.metadata.json
@@ -1,15 +1,15 @@
 {
   "Provider": "m365",
-  "CheckID": "defenderendpoint_privileged_user_exposed_credentials",
+  "CheckID": "defenderxdr_endpoint_privileged_user_exposed_credentials",
   "CheckTitle": "Ensure privileged users do not have credentials exposed on vulnerable endpoints",
   "CheckType": [],
-  "ServiceName": "defenderendpoint",
+  "ServiceName": "defenderxdr",
   "SubServiceName": "",
   "ResourceIdTemplate": "",
   "Severity": "critical",
   "ResourceType": "Exposure Management",
   "ResourceGroup": "security",
-  "Description": "Privileged users may have authentication artifacts (CLI secrets, cookies, tokens) exposed on endpoints with high risk scores. Microsoft Defender for Endpoint's Security Exposure Management detects when credentials from users with Entra ID privileged roles are present on vulnerable devices.",
+  "Description": "Privileged users may have authentication artifacts (CLI secrets, cookies, tokens) exposed on endpoints with high risk scores. Microsoft Defender XDR's Security Exposure Management detects when credentials from users with Entra ID privileged roles are present on vulnerable devices.",
   "Risk": "Exposed credentials on vulnerable endpoints enable account takeover through stolen tokens or cookies, Conditional Access bypass via primary refresh tokens, lateral movement to sensitive resources, and persistence until tokens are explicitly revoked.",
   "RelatedUrl": "",
   "AdditionalURLs": [
@@ -35,5 +35,5 @@
   ],
   "DependsOn": [],
   "RelatedTo": [],
-  "Notes": "This check requires Microsoft Defender for Endpoint with Security Exposure Management enabled. The ThreatHunting.Read.All permission is required to query the ExposureGraphEdges table via the Advanced Hunting API."
+  "Notes": "This check requires Microsoft Defender XDR with Security Exposure Management enabled. The ThreatHunting.Read.All permission is required to query the ExposureGraphEdges table via the Advanced Hunting API."
 }
