+++ prowler/providers/m365/services/defenderxdr/defenderxdr_endpoint_privileged_user_exposed_credentials/defenderxdr_endpoint_privileged_user_exposed_credentials.py
@@ -1,19 +1,19 @@
-"""Check for exposed credentials of privileged users in Defender for Endpoint.
+"""Check for exposed credentials of privileged users in Defender XDR.
 
 This check identifies privileged users whose authentication credentials
 (CLI secrets, cookies, tokens) are exposed on vulnerable endpoints.
 """
 
 from prowler.lib.check.models import Check, CheckReportM365
-from prowler.providers.m365.services.defenderendpoint.defenderendpoint_client import (
-    defenderendpoint_client,
+from prowler.providers.m365.services.defenderxdr.defenderxdr_client import (
+    defenderxdr_client,
 )
 
 
-class defenderendpoint_privileged_user_exposed_credentials(Check):
+class defenderxdr_endpoint_privileged_user_exposed_credentials(Check):
     """Check if privileged users have exposed credentials on endpoints.
 
-    This check queries Microsoft Defender for Endpoint's ExposureGraphEdges
+    This check queries Microsoft Defender XDR's ExposureGraphEdges
     table via the Advanced Hunting API to identify privileged users whose
     authentication artifacts (CLI secrets, user cookies, sensitive tokens)
     are exposed on endpoints with high risk or exposure scores.
@@ -36,19 +36,19 @@ class defenderendpoint_privileged_user_exposed_credentials(Check):
         findings = []
 
         # Step 1: Check MDE status
-        mde_status = defenderendpoint_client.mde_status
+        mde_status = defenderxdr_client.mde_status
 
         # API call failed - likely missing ThreatHunting.Read.All permission
         if mde_status is None:
             report = CheckReportM365(
                 metadata=self.metadata(),
                 resource={},
-                resource_name="Defender for Endpoint",
+                resource_name="Defender XDR",
                 resource_id="mdeStatus",
             )
             report.status = "FAIL"
             report.status_extended = (
-                "Unable to query Microsoft Defender for Endpoint status. "
+                "Unable to query Microsoft Defender XDR status. "
                 "Verify that ThreatHunting.Read.All permission is granted."
             )
             findings.append(report)
@@ -59,7 +59,7 @@ class defenderendpoint_privileged_user_exposed_credentials(Check):
             report = CheckReportM365(
                 metadata=self.metadata(),
                 resource={},
-                resource_name="Defender for Endpoint",
+                resource_name="Defender XDR",
                 resource_id="mdeStatus",
             )
             report.status = "FAIL"
@@ -76,7 +76,7 @@ class defenderendpoint_privileged_user_exposed_credentials(Check):
             report = CheckReportM365(
                 metadata=self.metadata(),
                 resource={},
-                resource_name="Defender for Endpoint",
+                resource_name="Defender XDR",
                 resource_id="mdeDevices",
             )
             report.status = "PASS"
@@ -89,16 +89,14 @@ class defenderendpoint_privileged_user_exposed_credentials(Check):
             return findings
 
         # Step 2: MDE is active with devices - check for exposed credentials
-        exposed_credentials = (
-            defenderendpoint_client.exposed_credentials_privileged_users
-        )
+        exposed_credentials = defenderxdr_client.exposed_credentials_privileged_users
 
         # API call failed for exposed credentials query
         if exposed_credentials is None:
             report = CheckReportM365(
                 metadata=self.metadata(),
                 resource={},
-                resource_name="Defender for Endpoint",
+                resource_name="Defender XDR",
                 resource_id="exposedCredentials",
             )
             report.status = "FAIL"
@@ -137,7 +135,7 @@ class defenderendpoint_privileged_user_exposed_credentials(Check):
             report = CheckReportM365(
                 metadata=self.metadata(),
                 resource={},
-                resource_name="Defender for Endpoint Exposure Management",
+                resource_name="Defender XDR Exposure Management",
                 resource_id="exposedCredentials",
             )
             report.status = "PASS"
